Add files via upload
The files uploaded here are most of those I have edited, fixing many of the problems that prevented the application from compiling - however, there still remains a few issues - most of them are warnings (e.g. '<function-style-cast>': truncation from 'Thing*' to 'UWORD') with there currently being the issue of dplay.lib - unfortunately, it appears this library is no longer supplied with Windows' SDK (since being replaced by other newer SDKs, like those for XBL, GFW, etc.) and finding that lib file doesn't seem likely at the moment.
Importantly, the system does contain a lot of features I didn't look through (there's a lot of stuff going on in this project) and whether or not the deprecated modules and features (namely those relying upon DirectPlay and its libraries) can be upgraded/modified to newer standards is something I'm not sure of - nonetheless, going about this should be easier for everyone if these changes made here make the components more "coherent" overall.
This is the first branch, as there were over 100 files that were edited and GitHub limits the number of files per branch upload to 100, so there will be a second branch containing the remaining files.